Move Solr listing field definitions to og.base.solr.fields:
They are hardly REST API specific, and will need to be used by browser views
as well, not just REST API services.

In addition, reformat them so that they are easier to read and edit.
