+++ opengever/api/listing.py
@@ -1,11 +1,10 @@
 from ftw.solr.query import escape
 from ftw.solr.query import make_path_filter
-from opengever.api.solr_query_service import REQUIRED_RESPONSE_FIELDS
 from opengever.api.solr_query_service import SolrFieldMapper
 from opengever.api.solr_query_service import SolrQueryBaseService
 from opengever.base.interfaces import ISearchSettings
+from opengever.base.solr.fields import REQUIRED_RESPONSE_FIELDS
 from opengever.dossier.indexers import ParticipationIndexHelper
-from opengever.propertysheets.definition import SolrDynamicField
 from plone.registry.interfaces import IRegistry
 from plone.uuid.interfaces import IUUID
 from Products.CMFPlone.utils import safe_unicode
